Make FramingRxFSM a template
